Move variable text anchor code to new file
